+++ packages/core-utils/test/fees/fees.spec.ts
@@ -1,6 +1,9 @@
 import { expect } from '../setup'
 import * as fees from '../../src/fees'
-import { BigNumber } from 'ethers'
+import { BigNumber, utils } from 'ethers'
+
+const hundredBillion = 10 ** 11
+const million = 10 ** 6
 
 describe('Fees', () => {
   it('should count zeros and ones', () => {
@@ -18,115 +21,102 @@ describe('Fees', () => {
     }
   })
 
-  describe('Round L1 Gas Price', () => {
-    const roundL1GasPriceTests = [
-      { input: 10, expect: 10 ** 8, name: 'simple' },
-      { input: 10 ** 8 + 1, expect: 2 * 10 ** 8, name: 'one-over' },
-      { input: 10 ** 8, expect: 10 ** 8, name: 'exact' },
-      { input: 10 ** 8 - 1, expect: 10 ** 8, name: 'one-under' },
-      { input: 3, expect: 10 ** 8, name: 'small' },
-      { input: 2, expect: 10 ** 8, name: 'two' },
-      { input: 1, expect: 10 ** 8, name: 'one' },
-      { input: 0, expect: 0, name: 'zero' },
-    ]
-
-    for (const test of roundL1GasPriceTests) {
-      it(`should pass for ${test.name} case`, () => {
-        const got = fees.roundL1GasPrice(test.input)
-        const expected = BigNumber.from(test.expect)
-        expect(got).to.deep.equal(expected)
-      })
-    }
-  })
-
-  describe('Round L2 Gas Price', () => {
-    const roundL2GasPriceTests = [
-      { input: 10, expect: 10 ** 8 + 1, name: 'simple' },
-      { input: 10 ** 8 + 2, expect: 2 * 10 ** 8 + 1, name: 'one-over' },
-      { input: 10 ** 8 + 1, expect: 10 ** 8 + 1, name: 'exact' },
-      { input: 10 ** 8, expect: 10 ** 8 + 1, name: 'one-under' },
-      { input: 3, expect: 10 ** 8 + 1, name: 'small' },
-      { input: 2, expect: 10 ** 8 + 1, name: 'two' },
-      { input: 1, expect: 10 ** 8 + 1, name: 'one' },
-      { input: 0, expect: 1, name: 'zero' },
-    ]
-
-    for (const test of roundL2GasPriceTests) {
-      it(`should pass for ${test.name} case`, () => {
-        const got = fees.roundL2GasPrice(test.input)
-        const expected = BigNumber.from(test.expect)
-        expect(got).to.deep.equal(expected)
-      })
-    }
-  })
-
   describe('Rollup Fees', () => {
     const rollupFeesTests = [
       {
         name: 'simple',
         dataLen: 10,
-        l1GasPrice: 100_000_000,
-        l2GasPrice: 100_000_001,
+        l1GasPrice: utils.parseUnits('1', 'gwei'),
+        l2GasPrice: utils.parseUnits('1', 'gwei'),
         l2GasLimit: 437118,
-        error: false,
+      },
+      {
+        name: 'small-gasprices-max-gaslimit',
+        dataLen: 10,
+        l1GasPrice: utils.parseUnits('1', 'wei'),
+        l2GasPrice: utils.parseUnits('1', 'wei'),
+        l2GasLimit: 0x4ffffff,
+      },
+      {
+        name: 'large-gasprices-max-gaslimit',
+        dataLen: 10,
+        l1GasPrice: utils.parseUnits('1', 'ether'),
+        l2GasPrice: utils.parseUnits('1', 'ether'),
+        l2GasLimit: 0x4ffffff,
+      },
+      {
+        name: 'small-gasprices-max-gaslimit',
+        dataLen: 10,
+        l1GasPrice: utils.parseUnits('1', 'ether'),
+        l2GasPrice: utils.parseUnits('1', 'ether'),
+        l2GasLimit: 1,
+      },
+      {
+        name: 'max-gas-limit',
+        dataLen: 10,
+        l1GasPrice: utils.parseUnits('5', 'ether'),
+        l2GasPrice: utils.parseUnits('5', 'ether'),
+        l2GasLimit: 99_970_000,
       },
       {
         name: 'zero-l2-gasprice',
         dataLen: 10,
-        l1GasPrice: 100_000_000,
+
+        l1GasPrice: hundredBillion,
         l2GasPrice: 0,
         l2GasLimit: 196205,
-        error: true,
+
       },
       {
         name: 'one-l2-gasprice',
         dataLen: 10,
-        l1GasPrice: 100_000_000,
+        l1GasPrice: hundredBillion,
         l2GasPrice: 1,
         l2GasLimit: 196205,
-        error: false,
       },
       {
         name: 'zero-l1-gasprice',
         dataLen: 10,
         l1GasPrice: 0,
-        l2GasPrice: 100_000_001,
+        l2GasPrice: hundredBillion,
         l2GasLimit: 196205,
-        error: false,
       },
       {
         name: 'one-l1-gasprice',
         dataLen: 10,
         l1GasPrice: 1,
-        l2GasPrice: 23254,
+        l2GasPrice: hundredBillion,
         l2GasLimit: 23255,
-        error: true,
+      },
+      {
+        name: 'zero-gasprices',
+        dataLen: 10,
+        l1GasPrice: 0,
+        l2GasPrice: 0,
+        l2GasLimit: 23255,
+      },
+      {
+        name: 'larger-divisor',
+        dataLen: 10,
+        l1GasPrice: 0,
+        l2GasLimit: 10,
+        l2GasPrice: 0,
       },
     ]
 
     for (const test of rollupFeesTests) {
       it(`should pass for ${test.name} case`, () => {
         const data = Buffer.alloc(test.dataLen)
+        const got = fees.TxGasLimit.encode({
+          data,
+          l1GasPrice: test.l1GasPrice,
+          l2GasPrice: test.l2GasPrice,
+          l2GasLimit: test.l2GasLimit,
+        })
 
-        let got
-        let err = false
-        try {
-          got = fees.L2GasLimit.encode({
-            data,
-            l1GasPrice: test.l1GasPrice,
-            l2GasPrice: test.l2GasPrice,
-            l2GasLimit: test.l2GasLimit,
-          })
-        } catch (e) {
-          err = true
-        }
-
-        expect(err).to.equal(test.error)
-
-        if (!err) {
-          const decoded = fees.L2GasLimit.decode(got)
-          expect(decoded).to.deep.eq(BigNumber.from(test.l2GasLimit))
-        }
+        const decoded = fees.TxGasLimit.decode(got)
+        const roundedL2GasLimit = fees.ceilmod(test.l2GasLimit, 10_000)
+        expect(decoded).to.deep.eq(roundedL2GasLimit)
       })
     }
   })