+++ packages/data-transport-layer/src/services/l1-ingestion/service.ts
@@ -3,6 +3,7 @@ import { fromHexString, EventArgsAddressSet } from '@eth-optimism/core-utils'
 import { BaseService } from '@eth-optimism/common-ts'
 import { JsonRpcProvider } from '@ethersproject/providers'
 import { LevelUp } from 'levelup'
+import { ethers, constants } from 'ethers'
 
 /* Imports: Internal */
 import { TransportDB, TransportDBMapHolder, TransportDBMap} from '../../db/transport-db'
@@ -18,7 +19,6 @@ import { handleEventsTransactionEnqueued } from './handlers/transaction-enqueued
 import { handleEventsSequencerBatchAppended } from './handlers/sequencer-batch-appended'
 import { handleEventsStateBatchAppended } from './handlers/state-batch-appended'
 import { L1DataTransportServiceOptions } from '../main/service'
-import { constants } from 'ethers'
 
 export interface L1IngestionServiceOptions
   extends L1DataTransportServiceOptions {
@@ -67,6 +67,7 @@ export class L1IngestionService extends BaseService<L1IngestionServiceOptions> {
     contracts: OptimismContracts
     l1RpcProvider: JsonRpcProvider
     startingL1BlockNumber: number
+    l2ChainId: number
   } = {} as any
 
   protected async _init(): Promise<void> {
@@ -115,6 +116,10 @@ export class L1IngestionService extends BaseService<L1IngestionServiceOptions> {
       this.state.l1RpcProvider,
       this.options.addressManager
     )
+
+    this.state.l2ChainId = ethers.BigNumber.from(
+      await this.state.contracts.OVM_ExecutionManager.ovmCHAINID()
+    ).toNumber()
 
     const startingL1BlockNumber = await this.state.db.getStartingL1Block()
     if (startingL1BlockNumber) {
@@ -298,7 +303,7 @@ export class L1IngestionService extends BaseService<L1IngestionServiceOptions> {
             event,
             this.state.l1RpcProvider
           )
-          const parsedEvent = await handlers.parseEvent(event, extraData)
+          const parsedEvent = await handlers.parseEvent(event, extraData, this.state.l2ChainId)
           
           // filter chainId
           var chainId = event.args._chainId.toNumber()