+++ packages/message-relayer/src/service.ts
@@ -5,8 +5,7 @@ import { MerkleTree } from 'merkletreejs'
 
 /* Imports: Internal */
 import { fromHexString, sleep } from '@eth-optimism/core-utils'
-import { BaseService } from '@eth-optimism/common-ts'
-import SpreadSheet from './spreadsheet'
+import { Logger, BaseService, Metrics } from '@eth-optimism/common-ts'
 
 import { loadContract, loadContractFromManager } from '@metis.io/contracts'
 import { StateRootBatchHeader, SentMessage, SentMessageProof } from './types'
@@ -43,9 +42,11 @@ interface MessageRelayerOptions {
   // Number of blocks within each getLogs query - max is 2000
   getLogsInterval?: number
 
-  // Append txs to a spreadsheet instead of submitting transactions
-  spreadsheetMode?: boolean
-  spreadsheet?: SpreadSheet
+  // A custom logger to transport logs via; default STDOUT
+  logger?: Logger
+
+  // A custom metrics tracker to manage metrics; default undefined
+  metrics?: Metrics
 }
 
 const optionSettings = {
@@ -56,16 +57,12 @@ const optionSettings = {
   l2BlockOffset: { default: 1 },
   l1StartOffset: { default: 0 },
   getLogsInterval: { default: 2000 },
-  spreadsheetMode: { default: false },
 }
 
 export class MessageRelayerService extends BaseService<MessageRelayerOptions> {
   constructor(options: MessageRelayerOptions) {
     super('Message_Relayer', options, optionSettings)
   }
-
-  protected spreadsheetMode: boolean
-  protected spreadsheet: SpreadSheet
 
   private state: {
     lastFinalizedTxHeight: number
@@ -87,7 +84,6 @@ export class MessageRelayerService extends BaseService<MessageRelayerOptions> {
       pollingInterval: this.options.pollingInterval,
       l2BlockOffset: this.options.l2BlockOffset,
       getLogsInterval: this.options.getLogsInterval,
-      spreadSheetMode: this.options.spreadsheetMode,
     })
     // Need to improve this, sorry.
     this.state = {} as any
@@ -144,10 +140,6 @@ export class MessageRelayerService extends BaseService<MessageRelayerOptions> {
 
     this.logger.info('Connected to all contracts.')
 
-    if (this.options.spreadsheetMode) {
-      this.logger.info('Running in spreadsheet mode')
-    }
-
     this.state.lastQueriedL1Block = this.options.l1StartOffset
     this.state.eventCache = []
 
@@ -175,15 +167,15 @@ export class MessageRelayerService extends BaseService<MessageRelayerOptions> {
           }
         }
 
-        this.logger.info('Checking for newly finalized transactions...')
+        //this.logger.info('Checking for newly finalized transactions...')
         if (
           !(await this._isTransactionFinalized(
             this.state.nextUnfinalizedTxHeight
           ))
         ) {
-          this.logger.info('Did not find any newly finalized transactions', {
-            retryAgainInS: Math.floor(this.options.pollingInterval / 1000),
-          })
+          //this.logger.info('Did not find any newly finalized transactions', {
+          //  retryAgainInS: Math.floor(this.options.pollingInterval / 1000),
+          //})
 
           continue
         }
@@ -310,10 +302,10 @@ export class MessageRelayerService extends BaseService<MessageRelayerOptions> {
       startingBlock < (await this.options.l1RpcProvider.getBlockNumber())
     ) {
       this.state.lastQueriedL1Block = startingBlock
-      this.logger.info('Querying events', {
-        startingBlock,
-        endBlock: startingBlock + this.options.getLogsInterval,
-      })
+      //this.logger.info('Querying events', {
+      //  startingBlock,
+      //  endBlock: startingBlock + this.options.getLogsInterval,
+      //})
 
       const events: ethers.Event[] = await this.state.OVM_StateCommitmentChain.queryFilter(
         this.state.OVM_StateCommitmentChain.filters.StateBatchAppended(),
@@ -339,11 +331,12 @@ export class MessageRelayerService extends BaseService<MessageRelayerOptions> {
     const transaction = await this.options.l1RpcProvider.getTransaction(
       event.transactionHash
     )
-    const x = this.state.OVM_StateCommitmentChain.interface.decodeFunctionData(
-        'appendStateBatchByChainId',
-        transaction.data
-    )
-    const stateRoots = x['_batch']
+    const stateRoots =
+      this.state.OVM_StateCommitmentChain.interface.decodeFunctionData(
+         'appendStateBatchByChainId',
+         transaction.data
+      )['_batch']
+
     return {
       batch: {
         batchIndex: event.args._batchIndex,
@@ -494,69 +487,12 @@ export class MessageRelayerService extends BaseService<MessageRelayerOptions> {
     message: SentMessage,
     proof: SentMessageProof
   ): Promise<void> {
-    if (this.options.spreadsheetMode) {
-      try {
-        await this.options.spreadsheet.addRow({
-          chainId: this.options.l2ChainId,
-          target: message.target,
-          sender: message.sender,
-          message: message.message,
-          messageNonce: message.messageNonce.toString(),
-          encodedMessage: message.encodedMessage,
-          encodedMessageHash: message.encodedMessageHash,
-          parentTransactionIndex: message.parentTransactionIndex,
-          parentTransactionHash: message.parentTransactionIndex,
-          stateRoot: proof.stateRoot,
-          batchIndex: proof.stateRootBatchHeader.batchIndex.toString(),
-          batchRoot: proof.stateRootBatchHeader.batchRoot,
-          batchSize: proof.stateRootBatchHeader.batchSize.toString(),
-          prevTotalElements: proof.stateRootBatchHeader.prevTotalElements.toString(),
-          extraData: proof.stateRootBatchHeader.extraData,
-          index: proof.stateRootProof.index,
-          siblings: proof.stateRootProof.siblings.join(','),
-          stateTrieWitness: proof.stateTrieWitness.toString('hex'),
-          storageTrieWitness: proof.storageTrieWitness.toString('hex'),
-        })
-        this.logger.info('Submitted relay message to spreadsheet')
-      } catch (e) {
-        this.logger.error('Cannot submit message to spreadsheet')
-        this.logger.error(e.message)
-      }
-    } else {
-      try {
-        this.logger.info(
-          'Dry-run, checking to make sure proof would succeed...'
-        )
+    try {
+      this.logger.info('Dry-run, checking to make sure proof would succeed...')
 
-        await this.state.OVM_L1CrossDomainMessenger.connect(
-          this.options.l1Wallet
-        ).callStatic.relayMessageViaChainId(
-          this.options.l2ChainId,
-          message.target,
-          message.sender,
-          message.message,
-          message.messageNonce,
-          proof,
-          {
-            gasLimit: this.options.relayGasLimit,
-          }
-        )
-
-        this.logger.info(
-          'Proof should succeed. Submitting for real this time...'
-        )
-      } catch (err) {
-        this.logger.error('Proof would fail, skipping', {
-          message: err.toString(),
-          stack: err.stack,
-          code: err.code,
-        })
-        return
-      }
-
-      const result = await this.state.OVM_L1CrossDomainMessenger.connect(
+      await this.state.OVM_L1CrossDomainMessenger.connect(
         this.options.l1Wallet
-      ).relayMessageViaChainId(
+      ).callStatic.relayMessageViaChainId(
         this.options.l2ChainId,
         message.target,
         message.sender,
@@ -568,29 +504,52 @@ export class MessageRelayerService extends BaseService<MessageRelayerOptions> {
         }
       )
 
-      this.logger.info('Relay message transaction sent', {
-        transactionHash: result,
+      this.logger.info('Proof should succeed. Submitting for real this time...')
+    } catch (err) {
+      this.logger.error('Proof would fail, skipping', {
+        message: err.toString(),
+        stack: err.stack,
+        code: err.code,
       })
-
-      try {
-        const receipt = await result.wait()
-
-        this.logger.info('Relay message included in block', {
-          transactionHash: receipt.transactionHash,
-          blockNumber: receipt.blockNumber,
-          gasUsed: receipt.gasUsed.toString(),
-          confirmations: receipt.confirmations,
-          status: receipt.status,
-        })
-      } catch (err) {
-        this.logger.error('Real relay attempt failed, skipping.', {
-          message: err.toString(),
-          stack: err.stack,
-          code: err.code,
-        })
-        return
-      }
-      this.logger.info('Message successfully relayed to Layer 1!')
+      return
     }
+
+    const result = await this.state.OVM_L1CrossDomainMessenger.connect(
+      this.options.l1Wallet
+    ).relayMessageViaChainId(
+      this.options.l2ChainId,
+      message.target,
+      message.sender,
+      message.message,
+      message.messageNonce,
+      proof,
+      {
+        gasLimit: this.options.relayGasLimit,
+      }
+    )
+
+    this.logger.info('Relay message transaction sent', {
+      transactionHash: result,
+    })
+
+    try {
+      const receipt = await result.wait()
+
+      this.logger.info('Relay message included in block', {
+        transactionHash: receipt.transactionHash,
+        blockNumber: receipt.blockNumber,
+        gasUsed: receipt.gasUsed.toString(),
+        confirmations: receipt.confirmations,
+        status: receipt.status,
+      })
+    } catch (err) {
+      this.logger.error('Real relay attempt failed, skipping.', {
+        message: err.toString(),
+        stack: err.stack,
+        code: err.code,
+      })
+      return
+    }
+    this.logger.info('Message successfully relayed to Layer 1!')
   }
 }