+++ packages/contracts/CHANGELOG.md
@@ -1,5 +1,32 @@
 # Changelog
 
+## 0.3.5
+
+### Patch Changes
+
+- 4e03f8a9: Update contracts README to add deploy instructions.
+- 8e2bfd07: Introduces the congestion price oracle contract
+- 245136f1: Minor change to how deploy.ts is invoked
+- Updated dependencies [a64f8161]
+- Updated dependencies [750a5021]
+- Updated dependencies [c2b6e14b]
+  - @eth-optimism/core-utils@0.4.5
+
+## 0.3.4
+
+### Patch Changes
+
+- 7bf5941: Remove colon names from filenames
+- Updated dependencies [f091e86]
+- Updated dependencies [f880479]
+  - @eth-optimism/core-utils@0.4.4
+
+## 0.3.3
+
+### Patch Changes
+
+- 5e5d4a1: Separates logic for getting state dumps and making state dumps so we can bundle for browser
+
 ## 0.3.2
 
 ### Patch Changes