+++ packages/core-utils/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @eth-optimism/core-utils
 
+## 0.4.5
+
+### Patch Changes
+
+- a64f8161: Implement the next fee spec in both geth and in core-utils
+- 750a5021: Delete dead transaction coders. These are no longer used now that RLP encoded transactions are used
+- c2b6e14b: Implement the latest fee spec such that the L2 gas limit is scaled and the tx.gasPrice/tx.gasLimit show correctly in metamask
+
+## 0.4.4
+
+### Patch Changes
+
+- f091e86: Have watcher correctly handle failed L1 => L2 messages
+- f880479: End to end fee integration with recoverable L2 gas limit
+
 ## 0.4.3
 
 ### Patch Changes